# doc/openapi.yml
openapi: 3.0.4
info:
  title: Task Management API
  version: 1.0.0
  description: |
    This API provides endpoints for managing users, teams, projects, and tasks.
  contact:
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
paths:
...
  /projects/:
    post:
      summary: Create a new project
      description: Create a new project with the provided details.
      tags:
        - Projects
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
            example:
              title: "New Web Application"
              description: "Developing a responsive web app"
              team_id: "xyz789-4567-890a-bcdef1234567"
              status: "on_hold"
              priority: 3
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              example:
                _links:
                  self:
                    href: /projects/proj123-4567-890a-bcdef1234567
                    method: GET
                    title: View this project
                  collection:
                    href: /projects
                    method: GET
                    title: All projects
                  create:
                    href: /projects
                    method: POST
                    title: Create a new project
                  update:
                    href: /projects/proj123-4567-890a-bcdef1234567
                    method: PUT
                    title: Update this project
                  delete:
                    href: /projects/proj123-4567-890a-bcdef1234567
                    method: DELETE
                    title: Delete this project
                  tasks:
                    href: /tasks?project_id=proj123-4567-890a-bcdef1234567
                    method: GET
                    title: Tasks in this project
                  team:
                    href: /teams/xyz789-4567-890a-bcdef1234567
                    method: GET
                    title: Team assigned to this project
                  root:
                    href: /
                    method: GET
                    title: API root
                project_id: proj123-4567-890a-bcdef1234567
                title: Testing Web Application
                description: Testing a responsive web app
                status: on_hold
                team_id: xyz789-4567-890a-bcdef1234567
                category_id: null
                deadline: null
        '404':
          description: Invalid input – team not found
          content:
            application/json:
              example:
                _links:
                  create:
                    href: "/projects"
                    method: "POST"
                    title: "Create a new project"
                  root:
                    href: "/"
                    method: "GET"
                    title: "API root"
                  self:
                    href: "/projects"
                    method: "GET"
                    title: "All projects"
                  tasks:
                    href: "/tasks"
                    method: "GET"
                    title: "All tasks"
                  teams:
                    href: "/teams"
                    method: "GET"
                    title: "All teams"
                  users:
                    href: "/users"
                    method: "GET"
                    title: "All users"
                message: "Error creating project: Team not found"
        '401':
          description: Unauthorized
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"

    get:
      summary: Get all projects
      description: Retrieve a list of all projects.
      tags:
        - Projects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              example:
                _links:
                  create:
                    href: /projects
                    method: POST
                    title: Create a new project
                  root:
                    href: /
                    method: GET
                    title: API root
                  self:
                    href: /projects
                    method: GET
                    title: All projects
                  tasks:
                    href: /tasks
                    method: GET
                    title: All tasks
                  teams:
                    href: /teams
                    method: GET
                    title: All teams
                  users:
                    href: /users
                    method: GET
                    title: All users
                projects:
                  - _links:
                      self: /projects/proj123-4567-890a-bcdef1234567
                      tasks: /tasks?project_id=proj123-4567-890a-bcdef1234567
                    project_id: proj123-4567-890a-bcdef1234567
                    title: Project Alpha
                    description: Frontend redesign and API integration
                    status: active
                    team_id: xyz789-4567-890a-bcdef1234567
                    category_id: null
                    deadline: null
                  - _links:
                      self: /projects/proj456-4567-890a-bcdef1234567
                      tasks: /tasks?project_id=proj456-4567-890a-bcdef1234567
                    project_id: proj456-4567-890a-bcdef1234567
                    title: Project Beta
                    description: Backend refactor and optimization
                    status: planning
                    team_id: xyz789-4567-890a-bcdef1234567
                    category_id: null
                    deadline: null
        '401':
          description: Unauthorized
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
  /projects/{project_id}:
    get:
      summary: Get project details
      description: Retrieve details of a specific project.
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "proj123-4567-890a-bcdef1234567"
      responses:
        '200':
          description: Project details
          content:
            application/json:
              example:
                _links:
                  collection:
                    href: /projects
                    method: GET
                    title: All projects
                  create:
                    href: /projects
                    method: POST
                    title: Create a new project
                  delete:
                    href: /projects/proj123-4567-890a-bcdef1234567
                    method: DELETE
                    title: Delete this project
                  root:
                    href: /
                    method: GET
                    title: API root
                  self:
                    href: /projects/proj123-4567-890a-bcdef1234567
                    method: GET
                    title: View this project
                  tasks:
                    href: /tasks?project_id=proj123-4567-890a-bcdef1234567
                    method: GET
                    title: Tasks in this project
                  team:
                    href: /teams/xyz789-4567-890a-bcdef1234567
                    method: GET
                    title: Team assigned to this project
                  update:
                    href: /projects/proj123-4567-890a-bcdef1234567
                    method: PUT
                    title: Update this project
                project_id: proj123-4567-890a-bcdef1234567
                title: New Web Application
                description: Developing a responsive web application for client X
                status: active
                category_id: null
                deadline: null
                team_id: xyz789-4567-890a-bcdef1234567
        '401':
          description: Unauthorized
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '404':
          description: Project not found
          content:
            application/json:
              example:
                _links:
                  create:
                    href: "/projects"
                    method: "POST"
                    title: "Create a new project"
                  root:
                    href: "/"
                    method: "GET"
                    title: "API root"
                  self:
                    href: "/projects"
                    method: "GET"
                    title: "All projects"
                  tasks:
                    href: "/tasks"
                    method: "GET"
                    title: "All tasks"
                  teams:
                    href: "/teams"
                    method: "GET"
                    title: "All teams"
                  users:
                    href: "/users"
                    method: "GET"
                    title: "All users"
                error: "Not Found"
                message: "Error retrieving project: Project with ID abc123-4567-890a-bcdef1234567 not found"


    put:
      summary: Update project details
      description: Update details of a specific project.
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "proj123-4567-890a-bcdef1234567"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
            example:
              title: "Updated Web Application"
              description: "Enhanced web app features"
      responses:
        '201':
          description: Project updated successfully
          content:
            application/json:
              example:
                _links:
                  self:
                    href: /projects/proj123-4567-890a-bcdef1234567
                    method: GET
                    title: View this project
                  collection:
                    href: /projects
                    method: GET
                    title: All projects
                  create:
                    href: /projects
                    method: POST
                    title: Create a new project
                  update:
                    href: /projects/proj123-4567-890a-bcdef1234567
                    method: PUT
                    title: Update this project
                  delete:
                    href: /projects/proj123-4567-890a-bcdef1234567
                    method: DELETE
                    title: Delete this project
                  tasks:
                    href: /tasks?project_id=proj123-4567-890a-bcdef1234567
                    method: GET
                    title: Tasks in this project
                  root:
                    href: /
                    method: GET
                    title: API root
                project_id: proj123-4567-890a-bcdef1234567
                title: Updated Web Application
                description: Enhanced web app features
                status: active
                team_id: null
                category_id: null
                deadline: null
        '400':
          description: Bad request
          content:
            application/json:
              example:
                error: "Bad Request"
                message: "Invalid team_id"
        '401':
          description: Unauthorized
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '404':
          description: Project not found
          content:
            application/json:
              example:
                _links:
                  create:
                    href: "/projects"
                    method: "POST"
                    title: "Create a new project"
                  root:
                    href: "/"
                    method: "GET"
                    title: "API root"
                  self:
                    href: "/projects"
                    method: "GET"
                    title: "All projects"
                  tasks:
                    href: "/tasks"
                    method: "GET"
                    title: "All tasks"
                  teams:
                    href: "/teams"
                    method: "GET"
                    title: "All teams"
                  users:
                    href: "/users"
                    method: "GET"
                    title: "All users"
                message: "Error retrieving project: Project with ID abc123-4567-890a-bcdef1234567 not found"


    delete:
      summary: Delete a project
      description: Delete a specific project.
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "proj123-4567-890a-bcdef1234567"
      responses:
        '200':
          description: Project deleted successfully
          content:
            application/json:
              example:
                _links:
                  create:
                    href: /projects
                    method: POST
                    title: Create a new project
                  root:
                    href: /
                    method: GET
                    title: API root
                  self:
                    href: /projects
                    method: GET
                    title: All projects
                  tasks:
                    href: /tasks
                    method: GET
                    title: All tasks
                  teams:
                    href: /teams
                    method: GET
                    title: All teams
                  users:
                    href: /users
                    method: GET
                    title: All users
                message: Project deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              example:
                msg: "Missing Authorization Header"
        '404':
          description: Project not found
          content:
            application/json:
              example:
                _links:
                  create:
                    href: "/projects"
                    method: "POST"
                    title: "Create a new project"
                  root:
                    href: "/"
                    method: "GET"
                    title: "API root"
                  self:
                    href: "/projects"
                    method: "GET"
                    title: "All projects"
                  tasks:
                    href: "/tasks"
                    method: "GET"
                    title: "All tasks"
                  teams:
                    href: "/teams"
                    method: "GET"
                    title: "All teams"
                  users:
                    href: "/users"
                    method: "GET"
                    title: "All users"
                message: "Error retrieving project: Project with ID abc123-4567-890a-bcdef1234567 not found"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProjectCreateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        team_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [planning, active, completed, on_hold, cancelled]
        priority:
          type: integer
          minimum: 1
          maximum: 3
      required:
        - title
        - team_id
        - status
        - priority
    ProjectUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        team_id:
          type: string
          format: uuid
...